creates requests module and request strategies
