+++ fedbiomed/researcher/requests/__init__.py
@@ -1,0 +1,14 @@
+from ._requests import Requests 
+from ._strategies import ContinueOnDisconnect, \
+    ContinueOnError, \
+    StopOnAnyDisconnect, \
+    StopOnAnyError
+
+__all__ = [
+    "Requests", 
+    "ContinueOnDisconnect",
+    "ContinueOnError",
+    "StopOnAnyDisconnect", 
+    "StopOnAnyError"
+]
+
